+++ src/sockaddr.rs
@@ -2,17 +2,19 @@ use std::mem::{self, size_of, MaybeUninit};
 use std::net::{SocketAddr, SocketAddrV4, SocketAddrV6};
 use std::{fmt, io};
 
+#[cfg(windows)]
+use windows_sys::Win32::Networking::WinSock::SOCKADDR_IN6_0;
+
 use crate::sys::{
     sa_family_t, sockaddr, sockaddr_in, sockaddr_in6, sockaddr_storage, socklen_t, AF_INET,
     AF_INET6,
 };
-#[cfg(windows)]
-use winapi::shared::ws2ipdef::SOCKADDR_IN6_LH_u;
 
 /// The address of a socket.
 ///
 /// `SockAddr`s may be constructed directly to and from the standard library
 /// [`SocketAddr`], [`SocketAddrV4`], and [`SocketAddrV6`] types.
+#[derive(Clone)]
 pub struct SockAddr {
     storage: sockaddr_storage,
     len: socklen_t,
@@ -183,7 +185,7 @@ impl SockAddr {
                 addr.sin6_scope_id,
                 #[cfg(windows)]
                 unsafe {
-                    *addr.u.sin6_scope_id()
+                    addr.Anonymous.sin6_scope_id
                 },
             )))
         } else {
@@ -224,7 +226,7 @@ impl From<SocketAddrV4> for SockAddr {
         let sockaddr_in = sockaddr_in {
             sin_family: AF_INET as sa_family_t,
             sin_port: addr.port().to_be(),
-            sin_addr: crate::sys::to_in_addr(&addr.ip()),
+            sin_addr: crate::sys::to_in_addr(addr.ip()),
             sin_zero: Default::default(),
             #[cfg(any(
                 target_os = "dragonfly",
@@ -249,13 +251,6 @@ impl From<SocketAddrV4> for SockAddr {
 
 impl From<SocketAddrV6> for SockAddr {
     fn from(addr: SocketAddrV6) -> SockAddr {
-        #[cfg(windows)]
-        let u = unsafe {
-            let mut u = mem::zeroed::<SOCKADDR_IN6_LH_u>();
-            *u.sin6_scope_id_mut() = addr.scope_id();
-            u
-        };
-
         let sockaddr_in6 = sockaddr_in6 {
             sin6_family: AF_INET6 as sa_family_t,
             sin6_port: addr.port().to_be(),
@@ -264,7 +259,9 @@ impl From<SocketAddrV6> for SockAddr {
             #[cfg(any(unix, target_env = "sgx"))]
             sin6_scope_id: addr.scope_id(),
             #[cfg(windows)]
-            u,
+            Anonymous: SOCKADDR_IN6_0 {
+                sin6_scope_id: addr.scope_id(),
+            },
             #[cfg(any(
                 target_os = "dragonfly",
                 target_os = "freebsd",