+++ CHANGELOG.md
@@ -1,3 +1,51 @@
+# 0.4.5
+
+## Changed
+
+* Replace `winapi` dependency with `windows-sys`
+  (https://github.com/rust-lang/socket2/pull/303).
+
+## Added
+
+* `Socket::join_ssm_v4` and `Socket::leave_ssm_v4`
+  (https://github.com/rust-lang/socket2/pull/298).
+* `Socket::set_recv_tos` and `Socket::recv_tos`
+  (https://github.com/rust-lang/socket2/pull/299).
+
+## Fixed
+
+* Fix OpenBSD build
+  (https://github.com/rust-lang/socket2/pull/291).
+* Fix OpenBSD build
+  (https://github.com/rust-lang/socket2/pull/291).
+
+# 0.4.4
+
+## Fixed
+
+* Libc v0.2.114 fixed an issue where `ip_mreqn` where was not defined for Linux
+  s390x.
+
+# 0.4.3 (yanked)
+
+## Added
+
+* `Socket::set_fib`: sets `SO_SETFIB` (https://github.com/rust-lang/socket2/pull/271).
+* `Socket::attach_filter`, `SO_ATTACH_FILTER` (https://github.com/rust-lang/socket2/commit/6601ed132b37d6e9d178b34918bfb0b236800232).
+* `Socket::detach_filter`, `SO_DETACH_FILTER` (https://github.com/rust-lang/socket2/commit/6601ed132b37d6e9d178b34918bfb0b236800232).
+* `Socket::{header_included, set_header_included}`: sets or gets `IP_HDRINCL`
+  (https://github.com/rust-lang/socket2/commit/f9e882ee53c0b4e89c5043b6d709af95c9db5599).
+* `Socket::{cork, set_cork}`: sets or gets `TCP_CORK`
+  (https://github.com/rust-lang/socket2/commit/50f31f18aac8fd6ef277df2906adeeed9fa391de).
+* `Socket::{quickack, set_quickack}`: sets or gets `TCP_QUICKACK`
+  (https://github.com/rust-lang/socket2/commit/849eee2abc5d5170d2d3bc635386a2ba13b04530).
+* `Socket::{thin_linear_timeouts, set_thin_linear_timeouts}`: sets or gets
+  `TCP_THIN_LINEAR_TIMEOUTS`
+  (https://github.com/rust-lang/socket2/commit/24c231ca463a17f51e53e7a554c7915a95bdbcc7).
+* `Socket::{join_multicast_v4_n, leave_multicast_v4_n}`: extends the existing
+  multicast API by allowing an index to be used (in addition to an address)
+  (https://github.com/rust-lang/socket2/commit/750f83618b967c620bbfdf6ca04de7362bdb42b5).
+
 # 0.4.2
 
 ## Added
@@ -142,7 +190,6 @@
 * Use correct inmutable references in `Socket::send_with_flags` and
   `Socket::send_out_of_band`.
 * Use `IPPROTO_IPV6` in `Socket::join_multicast_v6` on Windows.
-* Don't assume the memory layout of `std::net::SocketAddr`.
 * Use `c_int` instead of `i32` where appropriate.
 
 ## From v0.4.0-alpha.1 to v0.4.0-alpha.2
@@ -163,3 +210,8 @@
 ## From v0.4.0-alpha.5 to v0.4.0
 
 * Use `SO_LINGER_SEC` on macOS for `Socket::get/set_linger`.
+
+# 0.3.16
+
+* Don't assume the memory layout of `std::net::SocketAddr`.
+* Other changes omited